% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate_equivalence.R
\name{deduplicate_equivalence}
\alias{deduplicate_equivalence}
\title{Deduplicatin using equivalence groups}
\usage{
deduplicate_equivalence(pairs, var = "duplicate_groups", selection, x)
}
\description{
Deduplicatin using equivalence groups
}

