% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_em.R
\name{fs_em}
\alias{fs_em}
\title{Calculate EM-estimates of m- and u-probabilities}
\usage{
fs_em(patterns, mprobs0 = list(0.95), uprobs0 = list(0.02), p0 = 0.05,
  tol = 1e-05)
}
\arguments{
\item{patterns}{either a table of patterns (as output by 
\code{\link{tabulate}}) or pairs with comparison columns (as output by
\code{\link{compare}}).}

\item{mprobs0, uprobs0}{initial values of the m- and u-probabilities. These
should be lists with numeric values. The names of the elements in the list
should correspond to the names in \code{by_x} in \code{\link{link}}.}

\item{p0}{the initial estimate of the probability that a pair is a match.}

\item{tol}{when the change in the m and u-probabilities is smaller than tol
the algorithm is stopped.}
}
\description{
Calculate EM-estimates of m- and u-probabilities
}

