% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linkexample1}
\alias{linkexample1}
\alias{linkexample2}
\title{Tiny example dataset for probabilistic linkage}
\format{Two data frames with resp. 6 and 5 records and 6 columns.}
\description{
Contains fictional records of 7 persons.
}
\details{
\itemize{
  \item \code{id} the id of the person; this contains no errors and can be used to 
    validate the linkage. 
  \item \code{lastname} the last name of the person; contains errors.
  \item \code{firstname} the first name of the persons; contains errors.
  \item \code{address} the address; contains errors.
  \item \code{sex} the sex; contains errors and missing values.
  \item \code{postcode} the postcode; contains no errors. 
}
}
\keyword{datasets}
