% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_filter_for_deduplication.R
\name{pairs_filter_for_deduplication}
\alias{pairs_filter_for_deduplication}
\title{Remove pairs which do not have to be compared for deduplication}
\usage{
pairs_filter_for_deduplication(pairs)
}
\arguments{
\item{pairs}{a \code{pairs} object, such as generated by 
\code{\link{pairs_blocking}}}
}
\description{
In case of deduplication one tries to link a data set to itself. Therefore, 
comparisons only have to be made for records for which the index of the
records from the first data set is larger than the index fron the record from
the second data set.
}

