% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problink_em.R
\name{summary.problink_em}
\alias{summary.problink_em}
\title{Summarise the results from \code{\link{problink_em}}}
\usage{
\method{summary}{problink_em}(object, ...)
}
\arguments{
\item{object}{the \code{\link{problink_em}} object.}

\item{...}{ignored;}
}
\description{
Summarise the results from \code{\link{problink_em}}
}

