% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_em.R
\name{summary.simple_em}
\alias{summary.simple_em}
\title{Summarise the results from \code{\link{simple_em}}}
\usage{
\method{summary}{simple_em}(object, ...)
}
\arguments{
\item{object}{the \code{\link{simple_em}} object.}

\item{...}{ignored;}
}
\description{
Summarise the results from \code{\link{simple_em}}
}
