% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{town_names}
\alias{town_names}
\title{Spelling variations of a set of town names}
\format{Data frames with 584 records and two columns.}
\description{
Contains spelling variations found in various files of a set of town/village
names. Names were selected that contain 'rdam' or 'rdm'. The correct/official
names are also given. This data set can be used as an example data set for 
deduplication
}
\details{
\itemize{
  \item name the name of the town/village as found in the files
  \item official_name the official/correct name
}
}
\keyword{datasets}

