% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy}
\alias{greedy}
\title{Greedy one-to-one matching of pairs}
\usage{
greedy(x, y, weight)
}
\arguments{
\item{x}{id's of lhs of pairs}

\item{y}{id's of rhs of pairs}

\item{weight}{weight of pair}
}
\description{
Greedy one-to-one matching of pairs
}
\details{
Pairs with the highest weight are selected as long a neither the lhs as the
rhs are already selected in a pair with a higher weight.
}

