% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linkexample1}
\alias{linkexample1}
\alias{linkexample2}
\title{Tiny example dataset for probabilistic linkage}
\format{Two data frames with resp. 6 and 5 records and 6 columns.}
\description{
Contains fictional records of 7 persons.
}
\details{
\itemize{
  \item id the id of the person; this contains no errors and can be used to 
    validate the linkage. 
  \item lastname the lastname of the person; contains errors.
  \item firstname the firstname of the persons; contains errors.
  \item address the address; contains errors.
  \item sex the sex; contains errors and missing values.
  \item postcode the postcode; contains no errors. 
}
}
\keyword{datasets}

